requirements.txt
requests==2.31.0
python-dotenv==1.0.1
pyyaml==6.0.1
fhir.resources==7.1.0
fastapi==0.109.1
pydantic==2.10.5
pydantic-settings==2.7.1
uvicorn[standard]==0.34.0
